requirements-dev.txt
flake8==6.0.0
pytest-django==4.5.2
pytest-cov==4.0.0
pytest-factoryboy==2.2.1

-r requirements.txt